requirements.txt
pycryptodome
aiohttp